requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
defusedxml==0.7.1
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.61.0
fpdf2==2.8.5
greenlet==3.2.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.0.0
PyMySQL==1.1.2
pypdf==6.4.0
python-dotenv==1.2.1
pytz==2025.2
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
WTForms==3.2.1
